Add modal with IIIF viewer when interacting with image card
