+++ frontend/src/components/Image/IFrameTest.css
@@ -1,5 +1,5 @@
 .iiif-iframe {
-  padding-top: 125px;
-  width: 100vw;
   margin: 0 auto;
+  width: 85vw;
+  height: 85vh;
 }
